Creating a general userListTemplate for simplification
